made step size variable h double precision across the board and tuned h for some particular examples.  Also updated dd reference output for all examples.  Numbers were obtained on 32-bit linux installed on a 64-bit machine with uname "Linux cuban 2.6.22-16-generic #1 SMP i686 GNU/Linux".  The compiler was ifort 10.1 with arguments " -check all -g -O0 -traceback -fp-model strict"

+++ testRoundTrip/examples/GriewankRerouting/driver_scalar_adm_joint.f90
@@ -70,7 +70,7 @@ program driver
 
   double precision, dimension(2) :: x0
   double precision, dimension(2,2) :: res_adj
-  real :: h
+  double precision :: h
   integer :: n, m
   integer i,j,k
