+++ testRoundTrip/examples/GriewankRerouting/driver_scalar_tlm.f90
@@ -66,7 +66,7 @@ program driver
   type(active) y_ph1, y_ph2
 
   double precision, dimension(2) :: x0
-  real :: h
+  double precision :: h
   integer:: n,m
   integer i
